[1/2] Move methods from AzulClient to MirrorController
Breaks mypy
